requirements.txt
ollama
streamlit
faiss-cpu
sentence-transformers
PyPDF2
pymupdf
transformers
torch
numpy
pandas